refactor: split and organized AddressBook components
